requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
datasets==2.15.0
peft==0.7.1
bitsandbytes==0.41.3
sentencepiece==0.1.99
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0